chore: move graph to seperate module
